requirements.txt
reportlab==3.5.49
aiodns==2.0.0
aiohttp==3.6.2
aiofiles==0.5.0
fastapi==0.61.0
python-decouple==3.1
requests==2.22.0
ujson==1.35
uvicorn==0.10.4
python-multipart==0.0.5
pdfrw==0.4
pdfreader==0.1.2
email-validator==1.0.5
celery==4.4.7
SQLAlchemy==1.3.19
fastapi-utils==0.2.1
SQLAlchemy-Utils==0.35.0
boto3==1.14.51
psycopg2==2.8.3
flower==0.9.5
pypdf==3.1.0